+++ wagtailgeowidget/__init__.py
@@ -7,8 +7,8 @@ Wagtail-Geo-Widget is the complete map solution for your Wagtail site.
 """
 
 __title__ = "wagtailgeowidget"
-__version__ = "6.0.0"
-__build__ = 600
+__version__ = "6.1.0"
+__build__ = 610
 __author__ = "Martin Sandström"
 __license__ = "MIT"
 __copyright__ = "Copyright 2015-2022 Fröjd Interactive"